created base file
This script will fetch / unpack / install CSF Firewall and copy over custom files to the server.
It also has the GET Cluster IP at the top; this is for usage in csf.conf to add the IP's to CLUSTER_SEND
The process for new installers is check all IPs in the all host range (hosts) and add them to CLUSTER_SEND / CLUSTER_RSVD this way all new installs will add all existing servers. Code for CSF.conf is not included at this time.