v1
This is a script to insert during global runs; it will reboot the server and monitor it until it comes back online and then continue the play.
please be aware that changing the 5 min max and 90 wait a second delay is not recommended. I've had issues when playing around with that.